requirements.txt
opencv-python>=4.8.1.78,<=4.10.0.84
cmake>=3.22
pywaggle[all]==0.56.0
aiortc~=1.9.0
APScheduler>=3.10.1,<4.0.0
asyncua~=1.1.5
cachetools<6.0.0
cython~=3.0.0
python-dotenv~=1.0.0
fastapi>=0.100,<0.111
numpy>=2.0.0,<2.3.0
rsa<4.8,>=3.1.2
docutils<0.17,>=0.10
google-auth<3.0dev,>=2.26.1
opencv-contrib-python>=4.8.1.78,<=4.10.0.84
pillow>=11.0,<12.0
prometheus-fastapi-instrumentator<=6.0.0
redis~=5.0.0
google-resumable-media>=2.6.0
requests>=2.32.0,<3.0.0
rich>=13.0.0,<13.10.0
supervision>=0.25.1,<=0.30.0
pybase64~=1.0.0
scikit-image>=0.19.0,<=0.25.2
requests-toolbelt~=1.0.0
wheel>=0.38.1,<=0.45.1
setuptools>=70.0.0 
networkx>=3.0.0,<4.0.0
pydantic>=2.8.0,<2.12.0
pydantic-settings<2.8
openai>=1.12.0,<2.0.0
structlog>=24.1.0,<25.0.0
zxing-cpp~=2.2.0
proto-plus<2.0.0dev,>=1.22.3
googleapis-common-protos<2.0.dev0,>=1.56.2
protobuf!=3.20.0,!=3.20.1,!=4.21.0,!=4.21.1,!=4.21.2,!=4.21.3,!=4.21.4,!=4.21.5,<5.0.0.dev0,>=3.19.5
uritemplate<5,>=3.0.1
httplib2>=0.19.0
httplib2<1.dev0,>=0.19.0
botocore==1.31.54
boto3<=1.35.60
colorama<0.4.5
typing_extensions>=4.8.0,<=4.12.2
pydot~=2.0.0
shapely>=2.0.4,<2.1.0
tldextract~=5.1.2
packaging~=24.0
anthropic~=0.49.0
pandas>=2.2.3,<3.0.0
paho-mqtt~=1.6.1
pytest>=8.0.0,<9.0.0
tokenizers>=0.19.0,<0.22.0
slack-sdk~=3.33.4
twilio~=9.3.7
httpx~=0.28.1
pylogix==1.0.5
pymodbus>=3.6.9,<=3.8.3
backoff~=2.2.0
filelock>=3.12.0,<=3.17.0
rfdetr
pytz